Move statistic related code into a helper (#57)
* Move statistic related code into a helper

* pre-commit drives me crazy
