+++ custom_components/eyeonwater/sensor.py
@@ -5,12 +5,7 @@ from typing import Any
 
 from pyonwater import DataPoint, Meter
 
-from homeassistant.components.recorder import get_instance
-from homeassistant.components.recorder.models import StatisticData, StatisticMetaData
-from homeassistant.components.recorder.statistics import (
-    async_import_statistics,
-    get_last_statistics,
-)
+from homeassistant.components.recorder.statistics import async_import_statistics
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity
 from homeassistant.const import UnitOfTemperature
 from homeassistant.core import callback
@@ -19,72 +14,17 @@ from homeassistant.helpers.update_coordinator import (
     CoordinatorEntity,
     DataUpdateCoordinator,
 )
-from homeassistant.util import dt as dtutil
 
 from .const import DATA_COORDINATOR, DATA_SMART_METER, DOMAIN, WATER_METER_NAME
+from .statistic_helper import (
+    convert_statistic_data,
+    filter_newer_data,
+    get_last_imported_time,
+    get_statistic_metadata,
+)
 
 _LOGGER = logging.getLogger(__name__)
 _LOGGER.addHandler(logging.StreamHandler())
-
-
-def get_statistics_id(meter_id: str) -> str:
-    """Generate statistics ID for meter."""
-    return f"sensor.water_meter_{meter_id.lower()}"
-
-
-def get_statistic_metadata(meter: Meter) -> StatisticMetaData:
-    """Build statistic metadata for a given meter."""
-    name = f"{WATER_METER_NAME} {meter.meter_id}"
-    statistic_id = get_statistics_id(meter.meter_id)
-
-    return StatisticMetaData(
-        has_mean=False,
-        has_sum=True,
-        name=name,
-        source="recorder",
-        statistic_id=statistic_id,
-        unit_of_measurement=meter.native_unit_of_measurement,
-    )
-
-
-def convert_statistic_data(data: list[DataPoint]) -> list[StatisticData]:
-    """Convert statistics data to HA StatisticData format."""
-    return [
-        StatisticData(
-            start=row.dt,
-            sum=row.reading,
-            state=row.reading,
-        )
-        for row in data
-    ]
-
-
-async def get_last_imported_time(hass, meter):
-    """Return last imported data datetime."""
-    # https://github.com/home-assistant/core/blob/74e2d5c5c312cf3ba154b5206ceb19ba884c6fb4/homeassistant/components/tibber/sensor.py#L11
-
-    statistic_id = get_statistics_id(meter.meter_id)
-
-    last_stats = await get_instance(hass).async_add_executor_job(
-        get_last_statistics,
-        hass,
-        1,
-        statistic_id,
-        True,
-        {"start", "sum"},
-    )
-    _LOGGER.debug("last_stats %s", last_stats)
-
-    if last_stats:
-        date = last_stats[statistic_id][0]["start"]
-        _LOGGER.debug("date %d", date)
-        date = datetime.datetime.fromtimestamp(date, tz=dtutil.DEFAULT_TIME_ZONE)
-        _LOGGER.debug("date %d", date)
-        date = dtutil.as_local(date)
-        _LOGGER.debug("date %d", date)
-
-        return date
-    return None
 
 
 async def async_setup_entry(hass, config_entry, async_add_entities):
@@ -156,27 +96,12 @@ class EyeOnWaterSensor(CoordinatorEntity, SensorEntity):
         if self._available:
             self._state = self.meter.reading
 
-            self._last_historical_data = self.meter.last_historical_data.copy()
-            if self._last_imported_time and self._last_historical_data:
-                _LOGGER.info(
-                    "_last_imported_time %d - self._last_historical_data %d",
-                    self._last_imported_time,
-                    self._last_historical_data[-1].dt,
-                )
-                self._last_historical_data = list(
-                    filter(
-                        lambda r: r.dt > self._last_imported_time,
-                        self._last_historical_data,
-                    ),
-                )
-                _LOGGER.info(
-                    "%i data points will be imported",
-                    len(self._last_historical_data),
-                )
-
+            self._last_historical_data = filter_newer_data(
+                self.meter.last_historical_data,
+                self._last_imported_time,
+            )
             if self._last_historical_data:
                 self.import_historical_data()
-
                 self._last_imported_time = self._last_historical_data[-1].dt
 
         self.async_write_ha_state()
@@ -219,7 +144,7 @@ class EyeOnWaterTempSensor(CoordinatorEntity, SensorEntity):
 
     _attr_has_entity_name = True
     _attr_device_class = SensorDeviceClass.TEMPERATURE
-    _attr_native_unit_of_measurement=UnitOfTemperature.CELSIUS
+    _attr_native_unit_of_measurement = UnitOfTemperature.CELSIUS
 
     def __init__(
         self,
